fix: validate ePPN, handle whitespace-only ePPNs
Due to an issue with Django (https://github.com/clarin-eric/Centre-Registry/issues/6), a NULL ePPN value is not allowed. This has caused people to enter whitespace as ePPN. Whitespace strings are not the empty string or `None`, but are invalid in the Icinga configuration. A limited validation is now performed to handle this case.