submodules: Fix broken submodule handling
Previously, on the monitoring host, updates had the following steps:
1. Pull changes from `monitoring`
2. For every submodule:
   * Call `update(to_latest_revision = True)`
   * Checkout the main branch
3. Regenerate files based on the Centre Registry
4. Add, commit, and push files to Github

However, this was missing one crucial step: When the main branch of a
submodule is checked out, updates from its remote counterpart are not
yet synced and need to be pulled before proceeding any further.
Otherwise, we would add and commit changes from the regenerated files
to the _local_ main branch, thus diverging from the remote main branch.
It would then be impossible to push any of these changes.

To mitigate this issue, we modify step (2):
2. For every submodule:
   * Call `update()`
   * Checkout the main branch
   * Pull changes (at this point, we're synced)
   * Update `monitoring` to account for any submodule updates

With this change, there will additionally be no need to update
submodules manually in the future.